Renamed files under class names
